Basic Win32 port.
Visual Studio 6.0 C++ workspaces and projects are in the win32 subdirectory.

This is a CLIENT ONLY port.  Server functionality was not ported or tested.

Mechanism KERBEROS_V4 will work with kclient.

There are ReadMe.txt's in all projects, which describe what changes were made
to the default project wrapper, etc.