Encode the parameter values passed to auth_httpform, not the whole POST data
Before this change the whole application/x-www-form-urlencoded form data
was URL-%encoded. There was an exception made for "=" and "&" (as they
are special delimiters in this format), which resulted in the following problem:
any usernames/passwords/realm that contained "=" or "&" were not encoded
correctly. This patch fixes that by encoding only username/password/realm
values.

Note that this patch changes the earlier behavior, as remaining form data
is not encoded now.

The following issue is known with this change (it was present in the
original version of auth_httpform):

1) Spaces need to be replaced with "+". This needs to be done in usernames,
passwords and realms.

Cyrus SASL Bug # 3508

Based on patch by Jeroen van Meeuwen (Kolab Systems).