Don't overwrite PIC objects with non-PIC variant
This patch makes sure the non-PIC version of libsasldb.a, which
is created out of non-PIC objects, is not going to overwrite the PIC version,
which is created out of PIC objects. The PIC version is placed in .libs, and
the non-PIC version in the current directory.  This ensures that both non-PIC
and PIC versions are available in the correct locations.

Originally-by: Fabian Fagerholm <fabbe@debian.org>
Signed-off-by: Bastian Germann <bage@debian.org>