Fix for DoS attack from Oracle:
We have found some issues in saslauthd. We contacted various members who
maintain saslauthd and they suggested we report this to you so that
you can help to coordinate this issue with the saslauthd and other
consumers using this library. The details are below.

TLP:  AMBER

Libsasl2:  DOS attack possible on saslauthd deamon if configured with doors

Problem Description:

If "--with-ipctype=doors" is used, there are 2 types or stages of DOS
possible for saslauthd.  It may result in Segmentation Fault(core dump)
led by SIGSEGV, also it may result in leaking reference count under
certain circumstances.

CONFIGURE_OPTIONS += --with-ipctype=doors

The SIGSEGV can be triggered by running the PoC code as an unprivileged
user. There might be a similar case that could be created for the other
ipctype, this PoC only applies to the doors type.

When reference count is leaked but the SIGSEGV is prevented, it is
possible to lock up saslauthd daemon instead of killing it, which in a
way is worse, if there is a re-starter available.

The suggested fix addresses (1)the SIGSEGV, (2)the reference count
problem, and also (3) a mis-declaration issue which prevented compiling
the code on Solaris.

Using the PoC code (open-close-w-null.c) it was possible to prove that:

1.  The saslauthd can be DoS'ed by SIGSEGV.

2.  If SIGSEGV is prevented but the reference count link issue is not
fixed, it's possible to freeze saslauthd in a state where it has 1 or 0
threads and won't create any more because it thinks it already has the
maximum number of threads (usually 5).

There are a couple of curious things about the "simple-ugly" fix.  The
nice-complex fix keeps the thread count at maximum reference count,
which probably means it is a superior solution. However, that is for
upstream to ultimately decide.

A.  It's not clear why, with both fixes in (reference count and checking
the arguments) the thread gets killed, but it does NOT get killed if any
of the OTHER parameters that can be checked are
done wrong. For those cases, the thread appears to simply loop.

B. Sometimes saslauthd gets 1-off on the thread count, it shows 0
but still has 1 thread. But testing has not shown a way to get off by 2.

Both of these seem mild compared to the other 2 problems, but the
debugger didn't show exactly what is going on.

Both the simple-ugly fix and the nice-complex fix have been tested for
the core dump issue, and only the simple-ugly fix allows us to see the
reference-count issue, because it does not create a new thread until
there is a new connection request.

Our preference is for the project to release the fix by September 19, 2016.

TLP:  AMBER