Avoid read beyond end of buffer
The SASL client step 3 code in the SCRAM implementation is passed
buffer pointer and length parameters, but assumes that the buffer is
NUL-terminated when it uses strchr() to find the comma separator before
optional parameters. This can result in reading beyond the end of the
buffer. The code is changed to use memchr instead.

JIRA: OTHER-127

Test-information:

Bug detected via valgrind. No error after change and all seems to
work as expected.