windows build: delete unused '_sources' variables
These make the impression that changing them could have some effect.
In fact, this commit is already a victim of this:
  b179dc6e by Partha Susarla, 2017-07-25 06:14:40
  Move plugin_common.[ch] to a common subdir.

There, a rule for 'plugin_common.h' was added and it looked good because
'plugin_common.h' is listed in 'common_sources', but this in fact does
nothing and the added rule didn't do its job.

Signed-off-by: Alexandr Miloslavskiy <alexandr.miloslavskiy@syntevo.com>