Provide AUTHID to GSS to select correct principal
When using postfix smtp client with KRB5_CLIENT_KTNAME for relayhost
authentication, the first entry in the client keytab is always used.
For /etc/krb5.keytab, this might be a service principal of the host
instead of the user principal, so gss_init_sec_context then fails -
AS-REQ fails for service principals, only works for user principals.

Fix this by using the authid supplied by postfix to provide a
desired name to gss_acquire_cred (if present).

Postfix config:
  smtp_sasl_auth_enable = yes
  smtp_sasl_mechanism_filter = gssapi
  smtp_sasl_password_maps = static:HOSTNAME$$@EXAMPLE.ORG:irrelevant
  import_environment = ... KRB5_CLIENT_KTNAME=/etc/krb5.keytab \
                           KRB5CCNAME=MEMORY:

Signed-off-by: Michael Braun <michael-dev@fami-braun.de>